Removed sensor calculations (done by server)

+++ co2_sensor/include/Wifi.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void wifi_init_espnow(void);
-void send_HTTPS(uint16_t *co2, float *temp, uint8_t *rh, size_t size);
+void send_HTTPS(uint16_t *co2, uint16_t *temp, uint16_t *rh, size_t size);
 void initialize_wifi();
